+++ NEWS.md
@@ -1,3 +1,8 @@
+* Update dataverse paths for cumulative
+
+# ccesMRPprep 0.1.15
+* Update dataverse paths for 2025 
+
 # ccesMRPprep 0.1.14
 
 * Add 2024 election data to cd_info_2024